remove order value from filters
