+++ app/models/queries/work_packages/filter/attachment_content_filter.rb
@@ -44,8 +44,4 @@ class Queries::WorkPackages::Filter::AttachmentContentFilter < Queries::WorkPack
   def search_column
     'fulltext'
   end
-
-  def order
-    8
-  end
 end
