+++ app/models/queries/work_packages/filter/attachment_file_name_filter.rb
@@ -45,10 +45,6 @@ class Queries::WorkPackages::Filter::AttachmentFileNameFilter < Queries::WorkPac
     'file'
   end
 
-  def order
-    8
-  end
-
   private
 
   def normalize_text(text)
