+++ app/models/queries/work_packages/filter/done_ratio_filter.rb
@@ -31,8 +31,4 @@ class Queries::WorkPackages::Filter::DoneRatioFilter < Queries::WorkPackages::Fi
   def type
     :integer
   end
-
-  def order
-    14
-  end
 end
