+++ app/models/queries/work_packages/filter/estimated_hours_filter.rb
@@ -32,8 +32,4 @@ class Queries::WorkPackages::Filter::EstimatedHoursFilter <
   def type
     :integer
   end
-
-  def order
-    13
-  end
 end
