+++ spec/models/queries/work_packages/filter/created_at_filter_spec.rb
@@ -30,7 +30,6 @@ require 'spec_helper'
 
 describe Queries::WorkPackages::Filter::CreatedAtFilter, type: :model do
   it_behaves_like 'basic query filter' do
-    let(:order) { 9 }
     let(:type) { :datetime_past }
     let(:class_key) { :created_at }
