+++ spec/models/queries/work_packages/filter/dates_interval_filter_spec.rb
@@ -30,7 +30,6 @@ require 'spec_helper'
 
 describe Queries::WorkPackages::Filter::DatesIntervalFilter, type: :model do
   it_behaves_like 'basic query filter' do
-    let(:order) { 12 }
     let(:type) { :date }
     let(:class_key) { :dates_interval }
