+++ spec/models/queries/work_packages/filter/done_ratio_filter_spec.rb
@@ -30,7 +30,6 @@ require 'spec_helper'
 
 describe Queries::WorkPackages::Filter::DoneRatioFilter, type: :model do
   it_behaves_like 'basic query filter' do
-    let(:order) { 14 }
     let(:type) { :integer }
     let(:class_key) { :done_ratio }
