+++ spec/models/queries/work_packages/filter/subproject_filter_spec.rb
@@ -30,7 +30,6 @@ require 'spec_helper'
 
 describe Queries::WorkPackages::Filter::SubprojectFilter, type: :model do
   it_behaves_like 'basic query filter' do
-    let(:order) { 13 }
     let(:type) { :list }
     let(:class_key) { :subproject_id }
     let(:name) { I18n.t('query_fields.subproject_id') }
