+++ spec/models/queries/work_packages/filter/updated_at_filter_spec.rb
@@ -30,7 +30,6 @@ require 'spec_helper'
 
 describe Queries::WorkPackages::Filter::UpdatedAtFilter, type: :model do
   it_behaves_like 'basic query filter' do
-    let(:order) { 10 }
     let(:type) { :datetime_past }
     let(:class_key) { :updated_at }
